Docs: expand comments in AppDirectories
