+++ miataru/miataru/SettingsManagers/App Settings/Devices/DeviceLocationCacheStore.swift
@@ -88,15 +88,7 @@ class DeviceLocationCacheStore: ObservableObject {
     }
     
     private var fileURL: URL {
-        let fileManager = FileManager.default
-        let urls = fileManager.urls(for: .applicationSupportDirectory, in: .userDomainMask)
-        let appSupportURL = urls[0]
-        let bundleID = Bundle.main.bundleIdentifier ?? "DefaultApp"
-        let appDirectory = appSupportURL.appendingPathComponent(bundleID)
-        if !fileManager.fileExists(atPath: appDirectory.path) {
-            try? fileManager.createDirectory(at: appDirectory, withIntermediateDirectories: true, attributes: nil)
-        }
-        return appDirectory.appendingPathComponent(fileName)
+        AppDirectories.applicationSupportFile(named: fileName)
     }
     
     private init() {
